LC-find-the-smallest-divisor-given-a-threshold [Runtime: N/A; Memory: N/A]

+++ array/find-the-smallest-divisor-given-a-threshold.cpp
@@ -4,7 +4,7 @@ public:
         
         int thre = 0;
         for(int i:nums){
-            thre += ceil(i/n);
+            thre += ceil((double)i/n);
         }
         return thre;
     }
